Vendor indices based on our Dask contribution
This is just our code that we contributed to Dask. So there is no issue
with us vendoring it here. That said, we use a BSD 3-Clause license just
like Dask. So if there were any issue, we are basically using the same
license as well. This is just borrowed from the same vendoring in
`dask-ndmeasure`.

ref: https://github.com/dask-image/dask-ndmeasure/commit/3bfa65036a298242493619ea1b73dfc7d79c146e