Add compatibility module stub
Provides a compatibility module stub for containing various functions to
ensure compatibility with the much older Dask 0.13.0. This is needed as
this is the newest version of Dask for Python 3.4 on conda-forge. Once
we drop Python 3.4, we can use a newer version of Dask and drop most if
not all of this module's contents.