Sum over first two axes in hamming only
As we exploring comparing more than one array pair at a time, we need to
be able to pass multiple axes through and only reduce some of them.
Hence we adjust the `sum` in `hamming` to only reduce along the first
two axes, which have been used to represent the True/False pair
combinations. By making this change, it should be a trivial matter to
extend `hamming` to performing multiple comparisons at a time.