Vary bool arrays' shapes and chunks in 2D tests
Previously the bool arrays' shapes and chunks were the same for both
arrays in the 2D tests. However this is a bit restrictive and it makes
it harder to tell if things are working correctly. So allow the shapes
and chunks of the two bool arrays to vary independently in the 2D tests.