Wrap long line in setup.py
The `description` line in `setup`'s keyword arguments was a little too
long. So the string was placed inside parentheses and dropped to the
next line indented. This fixes the Flake8 issue raised, but also keeps
things readable while doing so. Verified manually the `description`
could still be accessed fine.