Add pdist
Implements `pdist` in a similar way to SciPy's own function of the same
name. This includes returning a condensed array of distance pairs.

Under the hood, `pdist` calls `cdist` to generate the distance matrix.
Operates on this distance matrix after the fact to try and remove all
chunks that are totally redundant by making use of `triu`. However
chunks that overlap with the diagonal may perform some redundant
computations that are later chucked out. Reducing the chunk size along
the number of points should minimize this waste.