Support the minkowski metric in cdist and pdist
Makes sure that the minkowski metric can be selected when using either
`cdist` or `pdist`. Provide support for passing along `p` to the
`minkowski` function. Also set `p` to `2` if it is undefined. Provide
tests comparing the results from `cdist` and `pdist` when using the
minkowski metric to those from SciPy to make sure they are consistent.
Also tests the case when `p` is not defined for both `cdist` and
`pdist`.