Flatten unselected result and mask in pdist
Before applying the mask to select out the upper triangle in `pdist`,
make sure to flatten the array first. Otherwise each chunk will be
flattened and combined instead. This creates a result whose order is
chunk dependent and thus will not match with SciPy's result for
non-trivial numbers of points per chunk.

Note: Uses our internal `ravel` function due to limitations of the
implementation in Dask 0.13.0. Also this implementation of `ravel` is a
bit more performant.