Drop use of concatenate from ravel
The `atop`-based `ravel` implementation was previously using the
`concatenate` argument with `atop`. This apparently was pulling the
whole array into memory at once, which is probably a bad idea. To fix
that, we drop the `concatenate` argument. However that means we need to
make some other changes as well. So we `rechunk` beforehand such that
the last dimension is equivalent to the array's shape. This makes
`ravel` a trivial operation for each chunk.