In `pdist`, manually slice from `cdist`
Instead of using `triu` and masking out the upper triangle from the
flattened array, simply slice out portions corresponding to the upper
triangle and use `concatenate` to join them all together into the
result.

This confers a number of benefits. First that the shape of the
result is known instead of being unknown. Second no zero arrays are
generated in the process. Third compatibility functions related to
reshaping and generating indices can be dropped reducing the maintenance
overhead.

As far as optimizations go, this should be similar to `triu` as it
results in dropping out the same chunks that `triu` would have. For the
chunks that are dropped out completely, this strategy will be more
performant than `triu` as it will not generate zeros in there place. For
the chunks along the diagonal, this should have similar performance as
there will be some duplication of effort for these computations. However
slicing out the selections of interest should be a little more
performant than calling NumPy's `triu` due to using less memory and
potentially avoiding a copy.