Drop rechunking of cdist's result
There should be no need to rechunk `cdist`'s result to match the input
arrays. After all this chunking should just fall out from the operations
that we have done in `cdist`. The only way the chunks would not match is
if there was some rechunking done behind the scenes. If rechunking was
done behind the scenes, there probably was a reason for it and thus we
should leave the result's chunks alone.

This rechunking was likely added to guarantee that `pdist`'s
`triu`-based optimization strategy would work (as `triu` requires square
chunks). However we have dropped the `triu`-based optimization from
`pdist` in favor of slicing out the relevant pieces to keep ourselves.
So there is no requirement from `pdist` to have `cdist` rechunk the
result.  Further if we did readd some `triu`-based optimization to
`pdist`, it would be up to `pdist` to guarantee the chunking was
appropriate. So `pdist`'s former requirement should not constrain
`cdist`'s result's chunks.