Use irange in pdist
Was using the builtin `range` in `pdist`, however its behavior differs
between Python 2 and Python 3, which makes it more likely to introduce
bugs due to Python version incompatibilities. Not to mention potentially
different performance profiles for the different Python versions. To fix
this, use our compatible `irange`, which is iterable on Python 2 and
Python 3.