Test pdist with a single point
In the case of a single point, `pdist` should return an empty array as
computing the metric of any point with itself should get tossed. However
this currently results in an error as `concatenate` is left to work with
an empty `list` devoid of Dask Arrays. The result is an error is thrown
instead. So we add this test to demonstrate this failure case and
attempt to fix it.