Handle a single point in pdist
In the case of a single point, `pdist` should return an empty array as
it will not compute pairwise distance between a point and itself.
However this previously was raising an error instead. To fix this, we
catch the case when there is nothing for `concatenate` to work with and
replace `result` by an empty Dask Array of the right shape and type.