Optimize pdist's handling of diagonal chunks
Make use of `cdist` for computing the bulk of the results for `pdist`.
However drop all duplicate chunks on the opposite side of the diagonal.
Though keep all chunks on the right side of the diagonal. As for any
chunks that land on the diagonal, break them up into pieces that include
non-duplicated pairs and pass those through `cdist` as well. Take all of
these results and flatten them so that they can be concatenated into one
big result.