Recursively handle pdist's diagonal chunks
Make use of `cdist` for computing the bulk of the results for `pdist`.
However drop all duplicate chunks on the opposite side of the diagonal.
Though keep all chunks on the right side of the diagonal.

As for any chunks that straddle the diagonal, recursively break them up
into smaller pieces. If the pieces are on the right side of the
diagonal, they are trivially handled with `cdist`. If they are on or
beyond the diagonal, they are trivially dropped. If they still
land on the diagonal, repeat the process by calling into `pdist` again
until they are resolved one of these two ways.

Since `pdist` returns its results in vector form, the recursive portion
needs to make use of `squareform` to convert them back into square
matrices that can be more easily worked with. Though the results are
again unraveled according to the constraints of `pdist`. So the brief
restructuring with `squareform` is a mere convenience to allow recursive
calls of `pdist` to proceed without issues.